voice-lab/requirements.runtime.txt
# CosyVoice2 inference runtime only (no training stack)
fastapi==0.116.1
uvicorn==0.35.0
numpy==1.26.4
soundfile==0.13.1
torch==2.8.0
torchaudio==2.8.0
transformers==4.45.2
huggingface-hub==0.34.4
modelscope==1.34.0
omegaconf==2.3.0
hydra-core==1.3.2
HyperPyYAML==1.2.3
conformer==0.3.2
diffusers==0.29.0
openai-whisper==20231117
onnxruntime-gpu==1.18.0; platform_system == "Linux"
onnxruntime==1.18.0; platform_system != "Linux"
librosa==0.11.0
inflect==7.5.0
wget==3.2
tqdm>=4.66
